Rework Battle and Rotation Improvements
Rework Battle so that they work with text rotation instead of complex logic. This means you can now use battles with margin. Also adds the ability to rotate set symbols from the set symbol tab and you can set the rotation in the packs. This also fixes how art rotation is handled so that the art rotates from the center of the art rather than a corner of the canvas and you can now set art rotation in the packs via artBounds. This adjusts the set symbol placement for a couple packs to reflect the rotation changes.