Station Improvements
The goal for this PR is to fix the Station frame issues some users were having where they would need to go into the ui tab and  press the reset button to get the station system working properly.
I moved all the pack specific Station setting into the version so the packs are less cluttered.
I made function loadScript a promise so that we don't ever have race conditions on larger scripts that do way more than most scripts. This version was having that issue on slower devices (tested on an old laptop and my phone over network, both had the issue at the top before this fix) and the eventual Minimalist frame also needs this as well to prevent loading issues for its version script. I figure this is a better option than picking an arbitrary amount of wait time and hope it works okay. This also prevents faster devices from being slowed down by such a delay.