Fixes classEdited Function
User reported this in the discord and I was able to reproduce the issue. When you would save a class card if it had an unused ability it would store the height for that ability as 1 in the save so it would get set to 2814 in the ui tab even though there should be no ability.

This fixes the default height value for the class abilities so that when you save a class card it doesn't save an unused ability (0 height) as a 1 and the abilities other than the first can have 0 values now.