Update packFCA.js (#20)
* Fix Outline On Text

Fix Outline on text so that it always is drawn first then

* Fixes outline color for mana symbols

* Fixed Outline on Text

This fixes how outline is drawn for text so that it is all drawn before the fill so we don't get overlap of outlines on previous lines of text. This also adjusts the Borderless Source Material Text so that it is closer in spacing now to accurate since we don't have to worry about text overlap with the outlines. Also fixes an issue we had with the replacement system for line breaks on the Borderless Source Material when {indent} was used and not ended beofre the next {lns} replacement

* Update packFCA.js

* Update packFCA.js

* Japan Showcase Fix

Fix Japan Showcase Nickname addon font name and add alt nickname to base Japan showcase that matches official style

* Update packFCA.js